Revert "expose `didToPublicKey` on `util`"
