+++ package.json
@@ -18,17 +18,13 @@
     "test-noble": "npx esbuild --platform=node --format=esm --bundle test/noble.ts | node --input-type=module | tap-spec",
     "test-sodium": "npm run test-sodium:node && npm run test-sodium:browser",
     "test:browser": "npm run build-tests:browser && npm run test-tape-run",
-    "test-sodium:node": "bash -c 'cd ./test && npx esbuild --platform=node --bundle sodium.ts | NODE_ENV=development node' | npx tap-spec",
+    "test-sodium:node": "bash -c 'cd ./test && npx esbuild --platform=node --bundle sodium.ts | NODE_ENV=development node' | tap-spec",
     "test-sodium:browser": "npm run build-tests-sodium:browser && npm run test-tape-run.sodium",
-    "test:node": "bash -c 'cd ./test && npx esbuild --format=esm --platform=node --bundle index.ts | node --input-type=module' | npx tap-spec",
-    "test-noble:node": "bash -c 'cd ./test && npx esbuild --platform=node --bundle noble.ts | NODE_ENV=development node' | npx tap-spec",
-    "test-noble:browser": "npm run build-tests-noble:browser && npm run test-tape-run.noble",
+    "test:node": "bash -c 'cd ./test && npx esbuild --format=esm --platform=node --bundle index.ts | node --input-type=module' | tap-spec",
     "build-tests:browser": "esbuild test/index.ts --target=es2020 --bundle --keep-names > test/test-bundle.js",
     "build-tests-sodium:browser": "esbuild test/sodium.ts --target=es2020 --bundle --keep-names > test/test-bundle.sodium.js",
-    "build-tests-noble:browser": "esbuild test/noble.ts --target=es2020 --bundle --keep-names > test/test-bundle.noble.js",
     "test-tape-run": "cat test/index.html | tape-run --input=html --static=test | tap-spec",
     "test-tape-run.sodium": "cat test/index.sodium.html | tape-run --input=html --static=test | tap-spec",
-    "test-tape-run.noble": "cat test/index.noble.html | tape-run --input=html --static=test | tap-spec",
     "build-cjs": "esbuild src/*.ts src/aes/*.ts src/ecc/*.ts src/rsa/*.ts ./src/sodium/*.ts --format=cjs --keep-names --tsconfig=tsconfig.build.json --outdir=./dist --out-extension:.js=.cjs --sourcemap",
     "build-esm": "esbuild src/*.ts src/aes/*.ts src/ecc/*.ts src/rsa/*.ts ./src/sodium/*.ts --format=esm --metafile=dist/meta.json --keep-names --tsconfig=tsconfig.build.json --outdir=./dist --sourcemap && tsc --emitDeclarationOnly --project tsconfig.build.json --outDir dist",
     "build-esm:min": "esbuild src/index.ts src/aes/*.ts src/ecc/*.ts src/rsa/*.ts ./src/sodium/*.ts --format=esm --keep-names --bundle --tsconfig=tsconfig.build.json --minify --out-extension:.js=.min.js --outdir=./dist --sourcemap",
