--- test/index.noble.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>tests</title>
-    <script src="/reporter.js"></script>
-</head>
-<body>
-    <script src="/test-bundle.noble.js"></script>
-</body>
-</html>
